FROMLIST: mm: Move the fallback definitions of pXXp_get()
Some platforms need to fix up the values when reading or writing page
tables. Because of this, the accessors must always be used; it is not
valid to simply dereference a pXX_t pointer.

Move these definitions up by a few lines, so they will be in scope
everywhere that currently dereferences a pXX_t pointer.

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>
Acked-by: David Hildenbrand (Red Hat) <david@kernel.org>
Link: https://lore.kernel.org/r/20251113014656.2605447-6-samuel.holland@sifive.com
Signed-off-by: Han Gao <gaohan@iscas.ac.cn>