FROMLIST: riscv: Fix logic for selecting DMA_DIRECT_REMAP
DMA_DIRECT_REMAP allows the kernel to make pages coherent for DMA by
remapping them in the page tables with a different pgprot_t value. On
RISC-V, this is supported by the page-based memory type extensions
(Svpbmt and Xtheadmae). It is independent from the software cache
maintenance extensions (Zicbom and Xtheadcmo).

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>
Link: https://lore.kernel.org/r/20251113014656.2605447-19-samuel.holland@sifive.com
Signed-off-by: Han Gao <gaohan@iscas.ac.cn>